DDF-3343 - Splits workspace model apart from workspace collection model
